git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/branches/logan@2981 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/dose/dose_dobexplorer.ss/src/graphwidget.cpp
@@ -55,11 +55,11 @@ void GraphWidget::paintEvent(QPaintEvent* event)
 
     QPainterPath path;
 
-    path.moveTo(start.secsTo(m_data.begin()->first), m_data.begin()->second);
+    path.moveTo(static_cast<qreal>(start.secsTo(m_data.begin()->first)), m_data.begin()->second);
     for (PlotData::iterator it = ++m_data.begin();
          it != m_data.end(); ++it)
     {
-        path.lineTo(start.secsTo(it->first),it->second);
+        path.lineTo(static_cast<qreal>(start.secsTo(it->first)),it->second);
     }
 
     painter.drawPath(path);
